requirements.txt
numpy
torch
PyYAML
matplotlib
tqdm
typing-extensions